Add ft_atof

+++ includes/libft.h
@@ -16,14 +16,14 @@
 # include "ft_type.h"
 # include <stdlib.h>
 
-int		ft_atof(const char *str);
+float	ft_atof(const char *str);
 int		ft_atoi(const char *str);
 int		ft_isalpha(int c);
 int		ft_isdigit(int c);
 int		ft_isalnum(int c);
 int		ft_isnumeric(char *str);
 int		ft_isascii(int c);
-int     ft_isspace(char c);
+int		ft_isspace(char c);
 void	*ft_memcpy(void *dst, const void *src, size_t n);
 void	*ft_memset(void *b, int c, size_t len);
 void	*ft_memmove(void *dst, const void *src, size_t len);
